api/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1
SQLAlchemy==2.0.41
Werkzeug==3.1.3 